+++ demo/admin/src/common/blocks/AccordionItemBlock.tsx
@@ -49,6 +49,7 @@ export const AccordionItemBlock = createCompositeBlock(
                     })),
                     required: true,
                 }),
+                hiddenInSubroute: true,
             },
             content: {
                 block: AccordionContentBlock,